--- scoperender.ts
@@ -1,3 +1,0 @@
-import { BasicREPL } from "./repl";
-import { Type, Value } from "./ast";
-import { NUM, BOOL, NONE, PyValue,CLASS } from "./utils";
